% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_disa_eid.R
\name{clean_disa_eid}
\alias{clean_disa_eid}
\title{Clean historic DISA DPI dataframe loaded from historic monthly data}
\usage{
clean_disa_eid(df)
}
\arguments{
\item{df}{Dataframe of historic monthly DISA DPI data before cleaning and join of metadata}
}
\value{
Final cleaned historic DISA DPI dataframe
}
\description{
Clean historic DISA DPI dataframe loaded from historic monthly data
}
\examples{
\dontrun{

 df <- clean_disa_eid()}
}
