% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_em.R
\name{clean_em}
\alias{clean_em}
\title{Process monthly enhanced monitoring submission datasets from PEPFAR Mozambique Clinical Partners}
\usage{
clean_em(df, type)
}
\arguments{
\item{df}{Processed monthly enhanced monitoring dataframe}

\item{type}{Submission tab}
}
\value{
A tidy dataframe with monthly enhanced monitoring results
}
\description{
Process monthly enhanced monitoring submission datasets from PEPFAR Mozambique Clinical Partners
}
\examples{
 \dontrun{

 df <- clean_em()}
}
