% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_em_ahd.R
\name{clean_em_ahd}
\alias{clean_em_ahd}
\title{Join AJUDA site metadata and clean processed monthly enhanced monitoring TB AHD dataframe}
\usage{
clean_em_ahd(df)
}
\arguments{
\item{df}{Processed monthly enhanced monitoring TB AHD dataframe}
}
\value{
Final TB AHD dataframe used in analytic dashboards
}
\description{
Join AJUDA site metadata and clean processed monthly enhanced monitoring TB AHD dataframe
}
\examples{
\dontrun{

df <- clean_em_ahd()}
}
