% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_em_ahdhiv.R
\name{clean_em_ahdhiv}
\alias{clean_em_ahdhiv}
\title{Join AJUDA site metadata and clean processed monthly enhanced monitoring AHD HIV dataframe}
\usage{
clean_em_ahdhiv(df, ajuda_site_map)
}
\arguments{
\item{df}{Processed monthly enhanced monitoring AHD HIV dataframe}

\item{ajuda_site_map}{Ajuda_site_map data}
}
\value{
Final AHD HIV dataframe used in analytic dashboards
}
\description{
Join AJUDA site metadata and clean processed monthly enhanced monitoring AHD HIV dataframe
}
\examples{
\dontrun{

df <- clean_em_ahdhiv()}
}
