% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_em_dsd.R
\name{clean_em_dsd}
\alias{clean_em_dsd}
\title{Join AJUDA site metadata and clean processed monthly enhanced monitoring DSD dataframe}
\usage{
clean_em_dsd(df)
}
\arguments{
\item{df}{Processed monthly enhanced monitoring DSD dataframe}
}
\value{
Final DSD dataframe used in analytic dashboards
}
\description{
Join AJUDA site metadata and clean processed monthly enhanced monitoring DSD dataframe
}
\examples{
\dontrun{

df <- clean_em_dsd()}
}
