% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_em_imer.R
\name{clean_em_imer}
\alias{clean_em_imer}
\title{Join AJUDA site metadata and clean processed monthly enhanced monitoring IMER dataframe}
\usage{
clean_em_imer(df)
}
\arguments{
\item{df}{Processed monthly enhanced monitoring IMER dataframe}
}
\value{
Final IMER dataframe used in analytic dashboards
}
\description{
Join AJUDA site metadata and clean processed monthly enhanced monitoring IMER dataframe
}
\examples{
\dontrun{

df <- clean_em_imer()}
}
