% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_em_mq.R
\name{clean_em_mq}
\alias{clean_em_mq}
\title{Join AJUDA site metadata and clean processed monthly enhanced monitoring MQ dataframe}
\usage{
clean_em_mq(df)
}
\arguments{
\item{df}{Processed monthly enhanced monitoring MQ dataframe}
}
\value{
Final MQ dataframe used in analytic dashboards
}
\description{
Join AJUDA site metadata and clean processed monthly enhanced monitoring MQ dataframe
}
\examples{
\dontrun{

df <- clean_em_mq()}
}
