% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_em_prep.R
\name{clean_em_prep}
\alias{clean_em_prep}
\title{Join AJUDA site metadata and clean processed monthly enhanced monitoring PrEP dataframe}
\usage{
clean_em_prep(df)
}
\arguments{
\item{df}{Processed monthly enhanced monitoring PrEP dataframe}
}
\value{
Final PrEP dataframe used in analytic dashboards
}
\description{
Join AJUDA site metadata and clean processed monthly enhanced monitoring PrEP dataframe
}
\examples{
\dontrun{

df <- clean_em_prep()}
}
