% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_em_tpt.R
\name{clean_em_tpt}
\alias{clean_em_tpt}
\title{Join AJUDA site metadata and clean processed monthly enhanced monitoring TPT dataframe}
\usage{
clean_em_tpt(df)
}
\arguments{
\item{df}{Processed monthly enhanced monitoring TPT dataframe}
}
\value{
Final TPT dataframe used in analytic dashboards
}
\description{
Join AJUDA site metadata and clean processed monthly enhanced monitoring TPT dataframe
}
\examples{
\dontrun{

df <- clean_em_tpt()}
}
