% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_em_txtb.R
\name{clean_em_txtb}
\alias{clean_em_txtb}
\title{Join AJUDA site metadata and clean processed monthly enhanced monitoring TXTB dataframe}
\usage{
clean_em_txtb(df)
}
\arguments{
\item{df}{Processed monthly enhanced monitoring TXTB dataframe}
}
\value{
Final TXTB dataframe used in analytic dashboards
}
\description{
Join AJUDA site metadata and clean processed monthly enhanced monitoring TXTB dataframe
}
\examples{
\dontrun{

df <- clean_em_txtb()}
}
