% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epi_model.R
\name{clean_mer}
\alias{clean_mer}
\title{A tibble with desired indicator grouped by PSNU and with age bands matching the age bands
in spectrum.}
\usage{
clean_mer(mer_data, indicator_value, standard_disag, num_dem)
}
\arguments{
\item{mer_data}{Data genie file in rds format}

\item{indicator_value}{one indicator from the data genie file}

\item{standard_disag}{relevant standarddisaggregates for the indicator}

\item{num_dem}{set to numerator as standard - numeratordenom field in datagenie}
}
\value{
A tibble with desired indicator grouped by PSNU and with age bands matching the age bands
in spectrum.
}
\description{
A tibble with desired indicator grouped by PSNU and with age bands matching the age bands
in spectrum.
}
\examples{
\dontrun{
   clean_mer(mer_data, "TX_CURR", "Age/Sex/HIVStatus", "N")
 }
}
