% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sdp.R
\name{clean_sdp}
\alias{clean_sdp}
\title{Clean SDP data from PSM}
\usage{
clean_sdp(SDP_data)
}
\arguments{
\item{SDP_data}{data from SPM relating to Supply Chain}
}
\value{
cleaned SDP data tibble
}
\description{
Clean SDP data from PSM
}
\examples{
  \dontrun{
   clean_sdp(sdp_data)
}
}
