% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epi_model.R
\name{create_epi_model}
\alias{create_epi_model}
\title{Creates a dataset for the desired indicator where all values are mapped by age group, sex and PSNU.
Any values assigned to the age group 15+ are divided in to the other age categories based on the proportion
in each district or province.}
\usage{
create_epi_model(mer_data, indicator_value, standard_disag, num_dem, label)
}
\arguments{
\item{mer_data}{Data genie file in rds format}

\item{indicator_value}{one indicator from the data genie file}

\item{standard_disag}{relevant standarddisaggregates for the indicator}

\item{num_dem}{set to numerator as standard - numeratordenom field in datagenie}

\item{label}{Text name of the indicator}
}
\value{
a tibble with the indicator broken down by PSNU, sex and age.
}
\description{
Creates a dataset for the desired indicator where all values are mapped by age group, sex and PSNU.
Any values assigned to the age group 15+ are divided in to the other age categories based on the proportion
in each district or province.
}
\examples{
 \dontrun{
   create_data(mer_data, "TX_CURR", "Age/Sex/HIVStatus", "N", "TX_CURR" )
 }
}
