% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_em_imer_var_map}
\alias{data_em_imer_var_map}
\title{Map for renaming variables contained in Enhanced Monitoring partner submission}
\format{
A mapping of variable names and recoded variable names
\describe{
\item{data_em_imer_var_map}{indicator}
\item{data_em_imer_var_map}{indicator_new}
}
}
\usage{
data(data_em_imer_var_map)
}
\description{
List of indicators
}
\keyword{datasets}
