% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sdp_orgunituid_map}
\alias{data_sdp_orgunituid_map}
\title{Map for renaming orgunituid contained in PSM SDP submittion}
\format{
A mapping of orgunituid and recoded orgunituid names
\describe{
\item{data_sdp_orgunituid_map}{orgunituid}
}
}
\usage{
data(data_sdp_orgunituid_map)
}
\description{
List of orgunituid
}
\keyword{datasets}
