% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_ats_autoteste}
\alias{data_sisma_ats_autoteste}
\title{SISMA ATS Auto Teste Map for Feature Engineering}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_sisma_smi_cpn_map}{indicator}
\item{data_sisma_smi_cpn_map}{indicator_new}
\item{data_sisma_smi_cpn_map}{age}
\item{data_sisma_smi_cpn_map}{sex}
\item{data_sisma_smi_cpn_map}{disaggregate}
}
}
\usage{
data(data_sisma_ats_autoteste)
}
\description{
List of indicators
}
\keyword{datasets}
