% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_ats_hist}
\alias{data_sisma_ats_hist}
\title{SISMA ATS Historic Map for Feature Engineering}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_sisma_smi_cpn_map}{indicator}
\item{data_sisma_smi_cpn_map}{indicator_new}
\item{data_sisma_smi_cpn_map}{modality}
\item{data_sisma_smi_cpn_map}{sub_group}
\item{data_sisma_smi_cpn_map}{result_status}
}
}
\usage{
data(data_sisma_ats_hist)
}
\description{
List of indicators
}
\keyword{datasets}
