% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_ats_results}
\alias{data_sisma_ats_results}
\title{SISMA ATS Results indicator Map for Feature Engineering}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_sisma_ats_results}{indicator}
\item{data_sisma_ats_results}{indicator_new}
\item{data_sisma_ats_results}{sex}
\item{data_sisma_ats_results}{age}
\item{data_sisma_ats_results}{age_coarse}
\item{data_sisma_ats_results}{modality}
\item{data_sisma_ats_results}{result_status}
\item{data_sisma_ats_results}{disaggregate}
}
}
\usage{
data(data_sisma_ats_results)
}
\description{
List of indicators
}
\keyword{datasets}
