% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_smi_cpn_map}
\alias{data_sisma_smi_cpn_map}
\title{Name of cascade plots available to the user}
\format{
A list of the plots available to the user
\describe{
\item{data_sisma_smi_cpn_map}{indicator}
\item{data_sisma_smi_cpn_map}{indicator_new}
\item{data_sisma_smi_cpn_map}{age}
\item{data_sisma_smi_cpn_map}{disaggregate}
}
}
\usage{
data(data_sisma_smi_cpn_map)
}
\description{
List of indicators
}
\keyword{datasets}
