% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epi_model.R
\name{epi_scenario_1}
\alias{epi_scenario_1}
\title{Creates a dataset for the indicator where the spectrum age bands are known. Any age bands without a value
are removed.}
\usage{
epi_scenario_1(indicator_data)
}
\arguments{
\item{indicator_data}{data for one indicator that has already been cleaned by clean_mer()}
}
\value{
A tibble with values for desired indicator for all ages except <15 and 15+
}
\description{
Creates a dataset for the indicator where the spectrum age bands are known. Any age bands without a value
are removed.
}
\examples{
 \dontrun{
   scenario_1(data_1)
 }
}
