% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_em_meta.R
\name{extract_em_meta}
\alias{extract_em_meta}
\title{Extract Metadata from monthly IP sumissions}
\usage{
extract_em_meta(filename, type)
}
\arguments{
\item{filename}{Local path to the monthly IP submission}

\item{type}{Meta element for extraction - set to either "ip" or "month"}
}
\value{
Meta values for ip and month
}
\description{
Extract Metadata from monthly IP sumissions
}
\examples{
\dontrun{

df <- extract_em_meta()}
}
