% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sisma_csv.R
\name{parse_sisma_csv}
\alias{parse_sisma_csv}
\title{Parse monthly sisma csv export}
\usage{
parse_sisma_csv(data, type)
}
\arguments{
\item{data}{Data input to be parsed}

\item{type}{Data type to be parsed}
}
\value{
A tidy dataframe with monthly sisma results
}
\description{
Parse monthly sisma csv export
}
\examples{
 \dontrun{

 df <- parse_sisma_csv()}
}
