% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_em_dsd.R
\name{plot_em_dsd}
\alias{plot_em_dsd}
\title{Generate summary table of DSD results for last 6 months}
\usage{
plot_em_dsd(df)
}
\arguments{
\item{df}{Cleaned DSD historic dataframe}
}
\value{
GT Table of 6 months indicator trends
}
\description{
Generate summary table of DSD results for last 6 months
}
\examples{
\dontrun{

df <- plot_em_dsd()}
}
