% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_em_imer.R
\name{plot_em_imer}
\alias{plot_em_imer}
\title{Generate summary table of IMER results for last 6 months}
\usage{
plot_em_imer(df)
}
\arguments{
\item{df}{Cleaned IMER historic dataframe}
}
\value{
GT Table of 6 months indicator trends
}
\description{
Generate summary table of IMER results for last 6 months
}
\examples{
\dontrun{

df <- plot_em_imer()}
}
