% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_em_mi.R
\name{plot_em_mi}
\alias{plot_em_mi}
\title{Generate summary table of MI results for last 6 months}
\usage{
plot_em_mi(df)
}
\arguments{
\item{df}{Cleaned MI historic dataframe}
}
\value{
GT Table of 6 months indicator trends
}
\description{
Generate summary table of MI results for last 6 months
}
\examples{
\dontrun{

df <- plot_em_mi()}
}
