% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_em_prep.R
\name{plot_em_prep}
\alias{plot_em_prep}
\title{Generate summary table of PrEP results for last 6 months}
\usage{
plot_em_prep(df)
}
\arguments{
\item{df}{Cleaned PrEP historic dataframe}
}
\value{
GT Table of 6 months indicator trends
}
\description{
Generate summary table of PrEP results for last 6 months
}
\examples{
\dontrun{

df <- plot_em_prep()}
}
