% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_em_tpt.R
\name{plot_em_tpt}
\alias{plot_em_tpt}
\title{Generate summary table of TPT results for last 6 months}
\usage{
plot_em_tpt(df)
}
\arguments{
\item{df}{Cleaned TPT historic dataframe}
}
\value{
GT Table of 6 months indicator trends
}
\description{
Generate summary table of TPT results for last 6 months
}
\examples{
\dontrun{

df <- plot_em_tpt()}
}
