% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_em_txtb.R
\name{plot_em_txtb}
\alias{plot_em_txtb}
\title{Generate summary table of TX_TB results for last 6 months}
\usage{
plot_em_txtb(df)
}
\arguments{
\item{df}{Cleaned TX_TB historic dataframe}
}
\value{
GT Table of 6 months indicator trends
}
\description{
Generate summary table of TX_TB results for last 6 months
}
\examples{
\dontrun{

df <- plot_em_txtb()}
}
