% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_em_prep.R
\name{prep_reshape}
\alias{prep_reshape}
\title{Process monthly enhanced monitoring PREP submission from PEPFAR Mozambique Clinical Partners}
\usage{
prep_reshape(filename, ip)
}
\arguments{
\item{filename}{Local path to the monthly IP submission}

\item{ip}{IP whose submission the file pertains to}
}
\value{
A tidy dataframe with monthly enhanced monitoring PREP results
}
\description{
Process monthly enhanced monitoring PREP submission from PEPFAR Mozambique Clinical Partners
}
\examples{
\dontrun{

df <- reshape_em_prep()}
}
