% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sisma_csv.R
\name{process_sisma_csv}
\alias{process_sisma_csv}
\title{Function that will clean sisma csv export and parse it into a tidy datafrom}
\usage{
process_sisma_csv(file, type)
}
\arguments{
\item{file}{Path of sisma csv input}

\item{type}{Type of sisma csv export (CPN, ATS-R, etc.)}
}
\value{
A tidy data frame of sisma program results
}
\description{
Function that will clean sisma csv export and parse it into a tidy datafrom
}
\examples{
 \dontrun{

 df <- process_sisma_csv()}
}
