% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_ajuda_sitemap.R
\name{pull_ajuda_sitemap}
\alias{pull_ajuda_sitemap}
\title{Retrieve PEPFAR Mozambique AJUDA Site Map from Google Sheets}
\usage{
pull_ajuda_sitemap()
}
\value{
A dataframe containing metadata for all PEPFAR Mozambique AJUDA sites
}
\description{
Retrieve PEPFAR Mozambique AJUDA Site Map from Google Sheets
}
\examples{
 \dontrun{

 df <- pull_ajuda_sitemap()
}

}
