% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_sitemap.R
\name{pull_sitemap}
\alias{pull_sitemap}
\title{Retrieve site meta data for Mozambique Health Facilities from the USAID Google Sheet}
\usage{
pull_sitemap(sheetname = "list_ajuda")
}
\arguments{
\item{sheetname}{Sheet name from the Mozambique AJUDA Site Map}
}
\value{
A dataframe containing metadata for all PEPFAR Mozambique AJUDA sites
}
\description{
Retrieve site meta data for Mozambique Health Facilities from the USAID Google Sheet
}
\examples{
 \dontrun{

 df <- pull_sitemap()
}

}
