% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_supplychain_metadata.R
\name{pull_supplychain_metadata}
\alias{pull_supplychain_metadata}
\title{Pull the supplychain metadata}
\usage{
pull_supplychain_metadata(sheetname = "product_metadata")
}
\arguments{
\item{sheetname}{name of the google sheet}
}
\value{
metadata
}
\description{
Pull the supplychain metadata
}
\examples{
  \dontrun{
   pull_supplychain_metadata(sheetname = "product_metadata")
}
}
