% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_disa_dpi.R
\name{reshape_disa_dpi}
\alias{reshape_disa_dpi}
\title{Reshape APHL DISA DPI submission}
\usage{
reshape_disa_dpi(filename, month)
}
\arguments{
\item{filename}{file path and name of APHL DPI submission}

\item{month}{Month for which DISA DPI is included in submission}
}
\value{
Tidy dataframe with APHL DPI results.  Does not include site metadata.
}
\description{
Reshape APHL DISA DPI submission
}
\examples{
\dontrun{

 df <- reshape_disa_dpi()}
}
