% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_disa_vl.R
\name{reshape_disa_vl}
\alias{reshape_disa_vl}
\title{Process monthly enhanced monitoring DISA Viral Load submission from PEPFAR Mozambique lab partner APHL}
\usage{
reshape_disa_vl(filepath, month)
}
\arguments{
\item{filepath}{Local path to the monthly DISA submission}

\item{month}{Month for which DISA Viral Load is included in submission}
}
\value{
A tidy dataframe with monthly enhanced monitoring DISA Viral Load results
}
\description{
Process monthly enhanced monitoring DISA Viral Load submission from PEPFAR Mozambique lab partner APHL
}
\examples{
 \dontrun{

 df <- reshape_disa_vl()}
}
