% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_em.R
\name{reshape_em}
\alias{reshape_em}
\title{Process monthly enhanced monitoring submission datasets from PEPFAR Mozambique Clinical Partners}
\usage{
reshape_em(type, filename, ip)
}
\arguments{
\item{type}{Submission tab}

\item{filename}{Local path to the monthly IP submission}

\item{ip}{IP whose submission the file pertains to}
}
\value{
A tidy dataframe with monthly enhanced monitoring results
}
\description{
Process monthly enhanced monitoring submission datasets from PEPFAR Mozambique Clinical Partners
}
\examples{
 \dontrun{

 df <- reshape_em()}
}
