% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_em_ahd.R
\name{reshape_em_ahd}
\alias{reshape_em_ahd}
\title{Process monthly enhanced monitoring TB AHD submission from PEPFAR Mozambique Clinical Partners}
\usage{
reshape_em_ahd(filename)
}
\arguments{
\item{filename}{Local path to the monthly IP submission}
}
\value{
A tidy dataframe with monthly TB AHD results.
}
\description{
Process monthly enhanced monitoring TB AHD submission from PEPFAR Mozambique Clinical Partners
}
\examples{
\dontrun{

df <- reshape_em_ahd()}
}
