% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_em_disa.R
\name{reshape_em_disa}
\alias{reshape_em_disa}
\title{Process monthly enhanced monitoring DISA submission from PEPFAR Mozambique lab partner APHL}
\usage{
reshape_em_disa(filename)
}
\arguments{
\item{filename}{Local path to the monthly DISA submission}
}
\value{
A tidy dataframe with monthly enhanced monitoring DISA results
}
\description{
Process monthly enhanced monitoring DISA submission from PEPFAR Mozambique lab partner APHL
}
\examples{
 \dontrun{

 df <- reshape_em_disa()}
}
