% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_em_prep.R
\name{reshape_em_prep}
\alias{reshape_em_prep}
\title{Process monthly enhanced monitoring PREP submission from PEPFAR Mozambique Clinical Partners}
\usage{
reshape_em_prep(filename)
}
\arguments{
\item{filename}{Local path to the monthly IP submission}
}
\value{
A tidy dataframe with monthly enhanced monitoring PREP results
}
\description{
Process monthly enhanced monitoring PREP submission from PEPFAR Mozambique Clinical Partners
}
\examples{
\dontrun{

df <- reshape_em_prep()}
}
