% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_em_tpt.R
\name{reshape_em_tpt}
\alias{reshape_em_tpt}
\title{Process monthly enhanced monitoring TPT submission from PEPFAR Mozambique Clinical Partners}
\usage{
reshape_em_tpt(filename, ip)
}
\arguments{
\item{filename}{Local path to the monthly IP submission}

\item{ip}{IP whose submission the file pertains to}
}
\value{
A tidy dataframe with monthly enhanced monitoring TPT results
}
\description{
Process monthly enhanced monitoring TPT submission from PEPFAR Mozambique Clinical Partners
}
\examples{
 \dontrun{

 df <- reshape_em_tpt()}
}
