% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_em_txtb.R
\name{reshape_em_txtb}
\alias{reshape_em_txtb}
\title{Process monthly enhanced monitoring TXTB submission from PEPFAR Mozambique Clinical Partners}
\usage{
reshape_em_txtb(filename)
}
\arguments{
\item{filename}{Local path to the monthly IP submission}
}
\value{
A tidy dataframe with monthly enhanced monitoring TXTB results
}
\description{
Process monthly enhanced monitoring TXTB submission from PEPFAR Mozambique Clinical Partners
}
\examples{
\dontrun{

df <- reshape_em_txtb()}
}
