% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_sup_kp.R
\name{reshape_sup_kp}
\alias{reshape_sup_kp}
\title{Process quarterly MER supplemental Key Population report}
\usage{
reshape_sup_kp(filename, ip)
}
\arguments{
\item{filename}{File containing quarterly supplemental MER results}

\item{ip}{Partner submitting quarterly results}
}
\value{
A tidy dataframe containing Key Population results - TX_PLVS, TX_RET, TX_CURR
}
\description{
Process quarterly MER supplemental Key Population report
}
\examples{
\dontrun{

df <- reshape_sup_kp()}
}
