% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_sup_trn.R
\name{reshape_sup_trn}
\alias{reshape_sup_trn}
\title{Process quarterly MER supplemental RTT and Transfer-in results}
\usage{
reshape_sup_trn(filename)
}
\arguments{
\item{filename}{File containing quarterly supplemental MER results}
}
\value{
A tidy dataframe containing supplemental RTT and TRNIF results
}
\description{
Process quarterly MER supplemental RTT and Transfer-in results
}
\examples{
\dontrun{

df <- reshape_sup_trn()}
}
