% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epi_model.R
\name{scenario_2}
\alias{scenario_2}
\title{Creates an estimate per age band for districts that report values for <15 or 15+.  The estimate is created using
the total values for each age band per district. (Scenario 2)}
\usage{
scenario_2(indicator_data)
}
\arguments{
\item{indicator_data}{data for one indicator that has already been cleaned by clean_mer()}
}
\value{
Tibble with an estimated value for any districts that use <15 or 15+ using PSNU proportion
}
\description{
Creates an estimate per age band for districts that report values for <15 or 15+.  The estimate is created using
the total values for each age band per district. (Scenario 2)
}
\examples{
\dontrun{
   scenario_2(data_2)
 }
}
