% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epi_model.R
\name{scenario_3}
\alias{scenario_3}
\title{Creates an estimate per age band for districts that report values for 15+.  The estimate is created using
the total values for each age band per province. (Scenario 3)}
\usage{
scenario_3(indicator_data)
}
\arguments{
\item{indicator_data}{data for one indicator that has already been cleaned by clean_mer()}
}
\value{
Tibble with an estimated value for any districts that use 15+ using SNU proportion
}
\description{
Creates an estimate per age band for districts that report values for 15+.  The estimate is created using
the total values for each age band per province. (Scenario 3)
}
\examples{
 \dontrun{
   scenario_3(data_3)
 }
}
