% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_em_meta.R
\name{extract_em_meta}
\alias{extract_em_meta}
\title{Process monthly enhanced monitoring IMER submission from PEPFAR Mozambique Clinical Partners}
\usage{
extract_em_meta(filename, type)
}
\arguments{
\item{filename}{Local path to the monthly IP submission}

\item{type}{Metadata type - ip or month}
}
\value{
Metadata values
}
\description{
Process monthly enhanced monitoring IMER submission from PEPFAR Mozambique Clinical Partners
}
\examples{
\dontrun{

df <- extract_em_meta()}
}
