% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_em_meta.R
\name{extract_em_meta}
\alias{extract_em_meta}
\title{Extract Partner and Period Metadata from monthly EM Excel submission}
\usage{
extract_em_meta(filename, type)
}
\arguments{
\item{filename}{Path to EM Excel submission}

\item{type}{Metadata to return.  Set to either "ip" or "month"}
}
\value{
values for Partner and Month
}
\description{
Extract Partner and Period Metadata from monthly EM Excel submission
}
\examples{
\dontrun{

df <- extract_em_meta()}
}
