% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_ajuda_sitemap.R
\name{pull_ajuda_sitemap}
\alias{pull_ajuda_sitemap}
\title{Import AJUDA site map}
\usage{
pull_ajuda_sitemap()
}
\value{
A dataframe of 640 rows for all AJUDA site metadata
}
\description{
Import AJUDA site map
}
\examples{
 \dontrun{
 site_map <- pull_ajuda_sitemap()
}

}
