% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_sitemap.R
\name{pull_sitemap}
\alias{pull_sitemap}
\title{Retreive site metadata}
\usage{
pull_sitemap(sheetname = "list_ajuda")
}
\arguments{
\item{sheetname}{Sheet name from the Mozambique AJUDA Site Map}
}
\value{
A dataframe containing site metadata
}
\description{
\code{pull_sitemap} utilizes stored user credentials to access
a GoogleSheet where Mozambique health facility metadata is
contained.  This includes site GIS coordinates, USG program
support definition, mapping of existing HIS, and unique system
identifiers
}
\examples{
 \dontrun{

 #fetch metadata for current PEPFAR AJUDA sites in Mozambique
 df <- pull_sitemap()

 #fetch DISA uid's mapped to Datim/Sisma uid
 df <- pull_sitemap(sheet = "map_disa")
}
}
