% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_sitemap.R
\name{pull_sitemap}
\alias{pull_sitemap}
\title{Retrieve PEPFAR Mozambique AJUDA Site Map from Google Sheets}
\usage{
pull_sitemap(sheetname = "list_ajuda")
}
\arguments{
\item{sheetname}{Sheet name from the Mozambique AJUDA Site Map}
}
\value{
A dataframe containing metadata for all PEPFAR Mozambique AJUDA sites
}
\description{
Retrieve PEPFAR Mozambique AJUDA Site Map from Google Sheets
}
\examples{
 \dontrun{

 df <- pull_sitemap()
}

}
