% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_em_dsd.R
\name{reshape_em_dsd}
\alias{reshape_em_dsd}
\title{Process monthly enhanced monitoring DSD submission from PEPFAR Mozambique Clinical Partners}
\usage{
reshape_em_dsd(filename)
}
\arguments{
\item{filename}{Local path to the monthly IP submission}
}
\value{
A tidy dataframe with monthly enhanced monitoring DSD results
}
\description{
Process monthly enhanced monitoring DSD submission from PEPFAR Mozambique Clinical Partners
}
\examples{
\dontrun{

df <- reshape_em_dsd()}
}
