% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_sup_vl.R
\name{reshape_sup_vl}
\alias{reshape_sup_vl}
\title{Process quarterly MER supplemental viral load source report}
\usage{
reshape_sup_vl(filename)
}
\arguments{
\item{filename}{File containing quarterly supplemental MER results}
}
\value{
A tidy dataframe containing TX_PVLS results according to the EPTS Ficha Mestra and EPTS Lab Module
}
\description{
Process quarterly MER supplemental viral load source report
}
\examples{
\dontrun{

df <- reshape_sup_vl()}
}
